Move development-only command test to new class
