+++ tests/LoadConfigurationsTest.php
@@ -51,19 +51,4 @@ final class LoadConfigurationsTest extends TestCase
     {
         $this->assertArrayNotHasKey('fake:removed', Artisan::all());
     }
-
-    public function testMakeTestCommandIsRemoved(): void
-    {
-        File::copy(base_path('production-config-app.php'), config_path('app.php'));
-
-        $this->refreshApplication();
-
-        $this->assertSame('production', $this->app->environment());
-        $this->assertArrayNotHasKey('test', Artisan::all());
-    }
-
-    public function tearDown(): void
-    {
-        File::copy(base_path('save-config-app.php'), config_path('app.php'));
-    }
 }
